change to int32_t

+++ src/pw_gv.h
@@ -16,18 +16,18 @@ namespace MyOpt {
   using Opts      = boost::program_options::options_description;
 }
 
-void printDist(std::ofstream &out, const std::vector<int> v, const std::string str, const long nread);
+void printDist(std::ofstream &out, const std::vector<int32_t> v, const std::string str, const uint64_t nread);
 
 class Fragment {
 public:
   std::string chr;
-  int F3;
+  int32_t F3;
   Strand strand;
-  int fraglen;
-  int readlen_F3;
+  int32_t fraglen;
+  int32_t readlen_F3;
 
  Fragment(): fraglen(0), readlen_F3(0) {}
- void addSAM(const std::vector<std::string> &v, const bool pair, const int sv) {
+ void addSAM(const std::vector<std::string> &v, const bool pair, const int32_t sv) {
    chr = rmchr(v[2]);
    readlen_F3 = v[9].length();
    if(pair) fraglen = abs(stoi(v[8]));
@@ -46,13 +46,13 @@ public:
 };
 
 class Read {
-  int weight;
+  int32_t weight;
   enum {WeightNum=1000};
  public:
-  int F3;
-  int F5;
-  int duplicate;
-  int inpeak;
+  int32_t F3;
+  int32_t F5;
+  int32_t duplicate;
+  int32_t inpeak;
   
  Read(const Fragment &frag): weight(WeightNum), F3(frag.F3), duplicate(0), inpeak(0) {
     if(frag.strand == STRAND_PLUS) F5 = frag.F3 + frag.fraglen;
@@ -67,9 +67,9 @@ class Read {
 class strandData {
  public:
   std::vector<Read> vRead;
-  long nread;
-  long nread_nonred;
-  long nread_red;
+  uint64_t nread;
+  uint64_t nread_nonred;
+  uint64_t nread_red;
   double nread_rpm;
   double nread_afterGC;
 
@@ -92,34 +92,34 @@ class strandData {
 
 class Wigstats {
   enum{n_mpDist=20, n_wigDist=200};
-  long sum;
+  uint64_t sum;
  public:
   double ave, var, nb_p, nb_n, nb_p0; // pois_p0
-  std::vector<long> mpDist;
-  std::vector<long> wigDist;
+  std::vector<uint64_t> mpDist;
+  std::vector<uint64_t> wigDist;
   std::vector<double> pwigDist;
 
- Wigstats(): sum(0),ave(0), var(0), nb_p(0), nb_n(0), nb_p0(0),
+ Wigstats(): sum(0), ave(0), var(0), nb_p(0), nb_n(0), nb_p0(0),
     mpDist(n_mpDist,0), wigDist(n_wigDist,0), pwigDist(n_wigDist,0) {}
 
-  double getPoisson(const int i) const {
+  double getPoisson(const int32_t i) const {
     if(ave) return _getPoisson(i, ave);
     else return 0;
   }
-  double getNegativeBinomial(const int i) const {
+  double getNegativeBinomial(const int32_t i) const {
     return _getNegativeBinomial(i, nb_p, nb_n);
   }
-  double getZINB(const int i) const {
+  double getZINB(const int32_t i) const {
     if(ave) return _getZINB(i, nb_p, nb_n, nb_p0);
     else return 0;
   }
-  int getwigDistthre() const {
-    int thre(9);
-    long num;
+  int32_t getwigDistthre() const {
+    int32_t thre(9);
+    uint64_t num;
     do{
       ++thre;
       num=0;
-      for(int i=0; i<thre; ++i) num += wigDist[i];
+      for(int32_t i=0; i<thre; ++i) num += wigDist[i];
     } while(num < sum*0.8 && thre <n_wigDist-1);
 #ifdef DEBUG
     BPRINT("\nthre %1%  (%2% / %3%)\n") % thre % num % sum;
@@ -127,31 +127,31 @@ class Wigstats {
     return thre;
   }
   void estimateParam() {
-    int thre = getwigDistthre();
+    int32_t thre = getwigDistthre();
     double par[thre+1];
     par[0] = thre;
-    for(int i=0; i<thre; ++i) par[i+1] = pwigDist[i];
+    for(int32_t i=0; i<thre; ++i) par[i+1] = pwigDist[i];
     iterateZINB(&par, nb_p, nb_n, nb_p, nb_n, nb_p0);
   }
   void setpWigDist() {
     for(size_t i=0; i<wigDist.size(); ++i) pwigDist[i] = wigDist[i]/static_cast<double>(sum);
   }
-  void getWigStats(const std::vector<int> &wigarray) {
-    int num95 = getPercentile(wigarray, 0.95);
+  void getWigStats(const std::vector<int32_t> &wigarray) {
+    int32_t num95 = getPercentile(wigarray, 0.95);
     
-    int size = wigDist.size();
-    std::vector<int> ar;
+    int32_t size = wigDist.size();
+    std::vector<int32_t> ar;
     for(auto x: wigarray) {
     if( x<0) std::cout << sum << "xxx" << x << std::endl;
     ++sum;
-      int v = WIGARRAY2VALUE(x);
+      int32_t v = WIGARRAY2VALUE(x);
       if(v < size) ++wigDist[v];
       if(x >= num95) continue;
       ar.push_back(v);
     }
     setpWigDist();
 
-    moment<int> mm(ar, 0);
+    moment<int32_t> mm(ar, 0);
     ave = mm.getmean();
     var = mm.getvar();
     nb_p = var ? ave/var : 0;
@@ -162,20 +162,20 @@ class Wigstats {
     //    std::cout << ave << "\t" << var << "\t" << nb_p << "\t" << nb_n<< std::endl;
     if(ave) estimateParam();
   }
-  void printwigDist(std::ofstream &out, const int i) const {
+  void printwigDist(std::ofstream &out, const int32_t i) const {
     out << boost::format("%1%\t%2%\t") % wigDist[i] % pwigDist[i];
   }
   void addmpDist(const double p) {
     if(!RANGE(p,0,1)) std::cout << "Warning: mappability " << p << " should be [0,1]" << std::endl;
-    else ++mpDist[(int)(p*n_mpDist)];
+    else ++mpDist[(int32_t)(p*n_mpDist)];
   }
   void addWigDist(const Wigstats &x) {
-    for(uint i=0; i<wigDist.size(); ++i) wigDist[i] += x.wigDist[i];
+    for(uint32_t i=0; i<wigDist.size(); ++i) wigDist[i] += x.wigDist[i];
     sum += x.sum;
     setpWigDist();
   }
   void printmpDist() const {
-    long num = accumulate(mpDist.begin(), mpDist.end(), 0);
+    uint64_t num = accumulate(mpDist.begin(), mpDist.end(), 0);
     for(size_t i=0; i<mpDist.size(); ++i)
       BPRINT("~%1%%%\t%2%\t%3%\n") % ((i+1)*100/mpDist.size()) % mpDist[i] % (mpDist[i]/static_cast<double>(num)); 
   }
@@ -190,44 +190,44 @@ class Wigstats {
 class SeqStats {
  protected:
   bool yeast;
-  long len, len_mpbl;
+  uint64_t len, len_mpbl;
   double weight4rpm;
   /* FRiP */
-  long nread_inbed;
+  uint64_t nread_inbed;
   //  GenomeCoverage gcov;
-  long nbp, ncov, ncovnorm;
+  uint64_t nbp, ncov, ncovnorm;
   double gcovRaw, gcovNorm;
  public:
   std::string name;
-  int nbin;
+  int32_t nbin;
 
   Wigstats ws;
   
   strandData seq[STRANDNUM];
   double depth;
   
- SeqStats(std::string s, int l=0): yeast(false), len(l), len_mpbl(l), weight4rpm(0), nread_inbed(0), nbp(0), ncov(0), ncovnorm(0), gcovRaw(0), gcovNorm(0) , nbin(0), depth(0) {
+ SeqStats(std::string s, int32_t l=0): yeast(false), len(l), len_mpbl(l), weight4rpm(0), nread_inbed(0), nbp(0), ncov(0), ncovnorm(0), gcovRaw(0), gcovNorm(0) , nbin(0), depth(0) {
     name = rmchr(s);
   }
   virtual ~SeqStats(){}
-  long getNreadInbed() const { return nread_inbed; }
+  uint64_t getNreadInbed() const { return nread_inbed; }
   void addfrag(const Fragment &frag) {
     Read r(frag);
     seq[frag.strand].vRead.push_back(r);
   }
-  long bothnread ()         const { return seq[STRAND_PLUS].nread         + seq[STRAND_MINUS].nread; }
-  long bothnread_nonred ()  const { return seq[STRAND_PLUS].nread_nonred  + seq[STRAND_MINUS].nread_nonred; }
-  long bothnread_red ()     const { return seq[STRAND_PLUS].nread_red     + seq[STRAND_MINUS].nread_red; }
-  long bothnread_rpm ()     const { return seq[STRAND_PLUS].nread_rpm     + seq[STRAND_MINUS].nread_rpm; }
-  long bothnread_afterGC () const { return seq[STRAND_PLUS].nread_afterGC + seq[STRAND_MINUS].nread_afterGC; }
+  uint64_t bothnread ()         const { return seq[STRAND_PLUS].nread         + seq[STRAND_MINUS].nread; }
+  uint64_t bothnread_nonred ()  const { return seq[STRAND_PLUS].nread_nonred  + seq[STRAND_MINUS].nread_nonred; }
+  uint64_t bothnread_red ()     const { return seq[STRAND_PLUS].nread_red     + seq[STRAND_MINUS].nread_red; }
+  uint64_t bothnread_rpm ()     const { return seq[STRAND_PLUS].nread_rpm     + seq[STRAND_MINUS].nread_rpm; }
+  uint64_t bothnread_afterGC () const { return seq[STRAND_PLUS].nread_afterGC + seq[STRAND_MINUS].nread_afterGC; }
 
-  long getlen()      const { return len; }
-  long getlenmpbl()  const { return len_mpbl; }
-  long getnbp()      const { return nbp; }
-  long getncov()     const { return ncov; }
-  long getncovnorm() const { return ncovnorm; }
+  uint64_t getlen()      const { return len; }
+  uint64_t getlenmpbl()  const { return len_mpbl; }
+  uint64_t getnbp()      const { return nbp; }
+  uint64_t getncov()     const { return ncov; }
+  uint64_t getncovnorm() const { return ncovnorm; }
   double getpmpbl()  const { return static_cast<double>(len_mpbl)/len; }
-  int getnbin()      const { return nbin; }
+  int32_t getnbin()      const { return nbin; }
   double getweight4rpm() const { return weight4rpm; }
   void print() const {
     std::cout << name << "\t" << len << "\t" << len_mpbl << "\t" << bothnread() << "\t" << bothnread_nonred() << "\t" << bothnread_red() << "\t" << bothnread_rpm() << "\t" << bothnread_afterGC()<< "\t" << depth << std::endl;
@@ -236,12 +236,12 @@ class SeqStats {
       if(lackOfRead4GenomeCov) out << boost::format("%1$.3f\t(%2$.3f)\t") % gcovRaw % gcovNorm;
       else out << boost::format("%1$.3f\t%2$.3f\t")   % gcovRaw % gcovNorm;
   }
-  void calcdepth(const int flen) {
+  void calcdepth(const int32_t flen) {
     depth = len_mpbl ? bothnread_nonred() * flen / static_cast<double>(len_mpbl): 0;
   }
-  void setF5(int flen) {
-    int d;
-    for(int strand=0; strand<STRANDNUM; ++strand) {
+  void setF5(int32_t flen) {
+    int32_t d;
+    for(int32_t strand=0; strand<STRANDNUM; ++strand) {
       if(strand == STRAND_PLUS) d = flen; else d = -flen;
       for(auto &x: seq[strand].vRead) x.F5 = x.F3 + d;
     }
@@ -251,7 +251,7 @@ class SeqStats {
   }
   void setWeight(const double weight) {
     weight4rpm = weight;
-    for(int i=0; i<STRANDNUM; i++) seq[i].nread_rpm = seq[i].nread_nonred * weight4rpm;
+    for(int32_t i=0; i<STRANDNUM; i++) seq[i].nread_rpm = seq[i].nread_nonred * weight4rpm;
   }
   void calcGcov(const std::vector<char> &array) {
     for(auto x:array) {
@@ -265,7 +265,7 @@ class SeqStats {
   void yeaston() { yeast = true; }
 
   bool isautosome() const {
-    int chrnum(0);
+    int32_t chrnum(0);
     try {
       chrnum = stoi(name);
     } catch (std::invalid_argument e) {  // 数値以外
@@ -303,7 +303,7 @@ class SeqStats {
 
 class SeqStatsGenome: public SeqStats {
   std::vector<bed> vbed;
-  void readGenomeTable(const std::string &, const int);
+  void readGenomeTable(const std::string &, const int32_t);
   
   void printstats() const {
     std::cout << "name\tlength\tlen_mpbl\tread num\tnonred num\tred num\tnormed\tafterGC\tdepth" << std::endl;
@@ -317,7 +317,7 @@ class SeqStatsGenome: public SeqStats {
   
  SeqStatsGenome(const MyOpt::Variables &values): SeqStats("Genome") {
     
-    readGenomeTable(values["gt"].as<std::string>(), values["binsize"].as<int>());
+    readGenomeTable(values["gt"].as<std::string>(), values["binsize"].as<int32_t>());
     if(values.count("mp")) getMpbl(values["mp"].as<std::string>(), chr);
     else if(values.count("mptable")) getMpbltable(values["mptable"].as<std::string>(), chr);
     for(auto &x:chr) {
@@ -331,12 +331,12 @@ class SeqStatsGenome: public SeqStats {
     for(auto &x:chr) if(yeast) x.yeaston();
     
     // sepchr
-    vsepchr = getVsepchr(values["threads"].as<int>());
+    vsepchr = getVsepchr(values["threads"].as<int32_t>());
     
 #ifdef DEBUG
     std::cout << "chr\tautosome" << std::endl;
     for(auto x:chr) std::cout << x.name << "\t" << x.isautosome() << std::endl;
-    for(uint i=0; i<vsepchr.size(); i++)
+    for(uint32_t i=0; i<vsepchr.size(); i++)
       std::cout << "thread " << (i+1) << ": "
 		<< vsepchr[i].s << "-" << vsepchr[i].e
 		<< std::endl;
@@ -344,10 +344,10 @@ class SeqStatsGenome: public SeqStats {
 #endif
   }
 
-  std::vector<sepchr> getVsepchr(const int);
+  std::vector<sepchr> getVsepchr(const int32_t);
   void setnread() {
     for(auto &x:chr) {
-      for(int i=0; i<STRANDNUM; i++) {
+      for(int32_t i=0; i<STRANDNUM; i++) {
 	x.seq[i].setnread();
 	seq[i].nread += x.seq[i].nread;
       }
@@ -355,7 +355,7 @@ class SeqStatsGenome: public SeqStats {
   }
   void setnread_red() {
     for(auto &x:chr) {
-      for(int i=0; i<STRANDNUM; i++) {
+      for(int32_t i=0; i<STRANDNUM; i++) {
 	seq[i].nread_nonred += x.seq[i].nread_nonred;
 	seq[i].nread_red    += x.seq[i].nread_red;
       }
@@ -363,7 +363,7 @@ class SeqStatsGenome: public SeqStats {
   }
   void setnread2nread_red() {
     for(auto &x:chr) {
-      for(int i=0; i<STRANDNUM; i++) x.seq[i].setnread_nonread_nofilter();
+      for(int32_t i=0; i<STRANDNUM; i++) x.seq[i].setnread_nonread_nofilter();
     }
   }
   void setbed(const std::string bedfilename) {
@@ -380,16 +380,16 @@ class SeqStatsGenome: public SeqStats {
     std::cout << "done." << std::endl;
     return;
   }
-   void calcdepthAll(const int flen) {
+   void calcdepthAll(const int32_t flen) {
      for (auto &x:chr) x.calcdepth(flen);
      calcdepth(flen);
   }
-   void setF5All(const int flen) {
+   void setF5All(const int32_t flen) {
     for (auto &x:chr) x.setF5(flen);
     setF5(flen);
   }
   std::vector<bed> getvbed() const { return vbed; }
-  void addGcov(const int i, boost::mutex &mtx) {
+  void addGcov(const int32_t i, boost::mutex &mtx) {
     boost::mutex::scoped_lock lock(mtx);
     nbp      += chr[i].getnbp();
     ncov     += chr[i].getncov();
@@ -401,22 +401,22 @@ class SeqStatsGenome: public SeqStats {
 
 class Mapfile {
   bool yeast;
-  const int ReadMax=200;
-  const int FragMax=1000;
-  int lenF3;
-  int lenF5;
-  int eflen;
-  int flen_def;
-  std::vector<int> vlenF3;
-  std::vector<int> vlenF5;
-  std::vector<int> vflen;
+  const int32_t ReadMax=200;
+  const int32_t FragMax=1000;
+  int32_t lenF3;
+  int32_t lenF5;
+  int32_t eflen;
+  int32_t flen_def;
+  std::vector<int32_t> vlenF3;
+  std::vector<int32_t> vlenF5;
+  std::vector<int32_t> vflen;
 
   std::string oprefix;
   std::string obinprefix;
   
   // PCR bias
-  int thre4filtering;
-  int nt_all, nt_nonred, nt_red;
+  int32_t thre4filtering;
+  int32_t nt_all, nt_nonred, nt_red;
   bool lackOfRead4Complexity;
   bool lackOfRead4GenomeCov;
   bool lackOfRead4FragmentVar;
@@ -424,7 +424,7 @@ class Mapfile {
   std::vector<Peak> vPeak;
 
   // GC bias
-  int maxGC;
+  int32_t maxGC;
 
   double backgroundUniformity;
 
@@ -433,17 +433,17 @@ class Mapfile {
   std::vector<SeqStats>::iterator lchr; // longest chromosome
 
   // Wigdist
-  int nwigdist;
-  std::vector<int> wigDist;
+  int32_t nwigdist;
+  std::vector<int32_t> wigDist;
   
  Mapfile(const MyOpt::Variables &values):
   yeast(false),
-    lenF3(0), lenF5(0), eflen(0), flen_def(values["flen"].as<int>()),
+    lenF3(0), lenF5(0), eflen(0), flen_def(values["flen"].as<int32_t>()),
     vlenF3(ReadMax,0), vlenF5(ReadMax,0), vflen(FragMax,0),
     thre4filtering(0), nt_all(0), nt_nonred(0), nt_red(0),
     lackOfRead4Complexity(false), lackOfRead4GenomeCov(false), lackOfRead4FragmentVar(false), r4cmp(0), maxGC(0), backgroundUniformity(0), genome(values)
     {
-      long lenmax(0);
+      uint64_t lenmax(0);
       for(auto itr = genome.chr.begin(); itr != genome.chr.end(); ++itr) {
 	if(lenmax < itr->getlenmpbl()) {
 	  lenmax = itr->getlenmpbl();
@@ -452,25 +452,25 @@ class Mapfile {
       }
 
       oprefix = values["odir"].as<std::string>() + "/" + values["output"].as<std::string>();
-      obinprefix = oprefix + "." + IntToString(values["binsize"].as<int>());
+      obinprefix = oprefix + "." + IntToString(values["binsize"].as<int32_t>());
     }
 
   void setbackgroundUniformity(const double bu) { backgroundUniformity = bu; }
   double getbackgroundUniformity() const { return backgroundUniformity; }
-  void setmaxGC(const int m) { maxGC = m; }
-  int getmaxGC() const {return maxGC; }
+  void setmaxGC(const int32_t m) { maxGC = m; }
+  int32_t getmaxGC() const {return maxGC; }
   void readGenomeTable(const MyOpt::Variables &values);
 
   void lackOfRead4Complexity_on() { lackOfRead4Complexity = true; }
   void lackOfRead4GenomeCov_on() { lackOfRead4GenomeCov = true; }
   void lackOfRead4FragmentVar_on() { lackOfRead4FragmentVar = true; }
-  int islackOfRead4GenomeCov() const { return lackOfRead4GenomeCov; };
+  int32_t islackOfRead4GenomeCov() const { return lackOfRead4GenomeCov; };
   void setthre4filtering(const MyOpt::Variables &values) {
-    if(values["thre_pb"].as<int>()) thre4filtering = values["thre_pb"].as<int>();
-    else thre4filtering = std::max(1, (int)(genome.bothnread() *10/static_cast<double>(genome.getlenmpbl())));
+    if(values["thre_pb"].as<int32_t>()) thre4filtering = values["thre_pb"].as<int32_t>();
+    else thre4filtering = std::max(1, (int32_t)(genome.bothnread() *10/static_cast<double>(genome.getlenmpbl())));
     std::cout << "Checking redundant reads: redundancy threshold " << thre4filtering << std::endl;
   }
-  int getthre4filtering() const { return thre4filtering; };
+  int32_t getthre4filtering() const { return thre4filtering; };
   void setr4cmp(const double r) { r4cmp = r; }
   double getr4cmp() const { return r4cmp; }
   void incNtAll() { ++nt_all; }
@@ -481,8 +481,8 @@ class Mapfile {
     std::ofstream out(filename);
 
     vPeak[0].printHead(out);
-    for(uint i=0; i<vPeak.size(); ++i) {
-      vPeak[i].print(out, i, values["binsize"].as<int>());
+    for(uint32_t i=0; i<vPeak.size(); ++i) {
+      vPeak[i].print(out, i, values["binsize"].as<int32_t>());
     }
   }
   std::string getprefix() const { return oprefix; }
@@ -498,11 +498,11 @@ class Mapfile {
     if(!values.count("nomodel")) out << "Estimated fragment length: " << eflen << std::endl;
     else out << "Predefined fragment length: " << flen_def << std::endl;
   }
-  void addF5(const int readlen_F5) { ++vlenF5[readlen_F5]; }
+  void addF5(const int32_t readlen_F5) { ++vlenF5[readlen_F5]; }
   void addfrag(const Fragment &frag) {
     ++vlenF3[frag.readlen_F3];
     ++vflen[frag.fraglen];
-    int on(0);
+    int8_t on(0);
     for(auto &x:genome.chr) {
       if(x.name == frag.chr) {
 	x.addfrag(frag);
@@ -562,10 +562,10 @@ class Mapfile {
     std::cout << "done." << std::endl;
     return;
     }*/
-  void seteflen(const int len) { eflen = len; }
-  int getlenF3() const { return lenF3; }
-  int getflen(const MyOpt::Variables &values) const {
-    int flen;
+  void seteflen(const int32_t len) { eflen = len; }
+  int32_t getlenF3() const { return lenF3; }
+  int32_t getflen(const MyOpt::Variables &values) const {
+    int32_t flen;
     if(!values.count("nomodel") || values.count("pair")) flen = eflen;
     else flen = flen_def;
     return flen;
@@ -573,15 +573,15 @@ class Mapfile {
   void addPeak(const Peak &peak) {
     vPeak.push_back(peak);
   }
-  void renewPeak(const int i, const double val, const double p) {
+  void renewPeak(const int32_t i, const double val, const double p) {
     vPeak[vPeak.size()-1].renew(i, val, p);
   }
 
   void estimateZINB() {
-    int thre = genome.ws.getwigDistthre();
+    int32_t thre = genome.ws.getwigDistthre();
     double par[thre+1];
     par[0] = thre;
-    for(int i=0; i<thre; ++i) par[i+1] = genome.ws.pwigDist[i];
+    for(int32_t i=0; i<thre; ++i) par[i+1] = genome.ws.pwigDist[i];
 
     //    iteratePoisson(&par, lchr->ave, genome.ave, genome.pois_p0);
     iterateZINB(&par, lchr->ws.nb_p, lchr->ws.nb_n, genome.ws.nb_p, genome.ws.nb_n, genome.ws.nb_p0);
